requirements.txt
pillow
PyMuPDF
opencv-python
pdf2image
httpx
PySide6
